Update Ubuntu installation

+++ core-install.sh
@@ -17,8 +17,7 @@ if [[ $(ls /etc/*release 2>/dev/null | grep 'redhat' 2>/dev/null) != '' ]]; then
   yum -y update
   yum -y groupinstall "Development tools"
   yum -y install openssl-devel readline-devel zlib-devel
-  yum -y install git unzip wget
-  yum -y install python-pip
+  yum -y install curl git unzip wget python-pip
 
   # Install Vagrant
   if [[ $(yum list | grep 'vagrant_1.8.1') == '' ]]; then
@@ -62,36 +61,56 @@ if [[ $(ls /etc/*release 2>/dev/null | grep 'redhat' 2>/dev/null) != '' ]]; then
 elif [[ $(ls /etc/*release 2>/dev/null | grep 'lsb' 2>/dev/null) != '' ]]; then
   echo "OS Detected: Debian Linux"
 
+  if [[ $(cat /etc/lsb-release | egrep ' 1[245]') == '' ]]; then
+    echo "ERROR: Unsupported Debian Linux version (only 12.x, 14.x and 15.x are currently supported)"
+    exit 1
+  fi
+
   # Install common packages
   apt-get -y update
   apt-get -y upgrade
-  apt-get -y install curl zlib1g-dev build-essential libssl-dev libreadline-dev libyaml-dev libsqlite3-dev sqlite3 libxml2-dev libxslt1-dev libcurl4-openssl-dev python-software-properties libffi-dev
-  apt-get -y install git-core unzip wget
+
+  if [[ $(cat /etc/lsb-release | egrep ' 1[45]') != '' ]]; then
+    apt-get -y install software-properties-common
+  else
+    apt-get -y install python-software-properties
+  fi
+
+  apt-get -y install build-essential curl sqlite3 git-core unzip wget
+  apt-get -y install zlib1g-dev libssl-dev libreadline-dev libyaml-dev libsqlite3-dev libxml2-dev libxslt1-dev libcurl4-openssl-dev libffi-dev libsdl1.2debian
 
   # Install Vagrant
-  wget https://releases.hashicorp.com/vagrant/1.8.1/vagrant_1.8.1_x86_64.deb
-  dpkg -i vagrant_1.8.1_x86_64.deb
-  rm vagrant_1.7.4_x86_64.deb
+  if [[ $(dpkg -l | egrep '^i' | egrep 'vagrant\s+1:1.8.1') == '' ]]; then
+    wget https://releases.hashicorp.com/vagrant/1.8.1/vagrant_1.8.1_x86_64.deb
+    dpkg -i vagrant_1.8.1_x86_64.deb
+    rm vagrant_1.8.1_x86_64.deb
+  else
+    echo 'Vagrant is already installed'
+  fi
 
   # Install VirtualBox
-  if [[ $(cat /etc/lsb-release | grep ' 15\.10') != '' ]]; then
-    wget http://download.virtualbox.org/virtualbox/5.0.8/virtualbox-5.0_5.0.8-103449~Ubuntu~wily_amd64.deb
-    dpkg -i virtualbox-5.0_5.0.8-103449~Ubuntu~wily_amd64.deb
-    rm virtualbox-5.0_5.0.8-103449~Ubuntu~wily_amd64.deb
-  elif [[ $(cat /etc/lsb-release | egrep ' (14|15\.04)') != '' ]]; then
-    wget http://download.virtualbox.org/virtualbox/5.0.8/virtualbox-5.0_5.0.8-103449~Ubuntu~trusty_amd64.deb
-    dpkg -i virtualbox-5.0_5.0.8-103449~Ubuntu~trusty_amd64.deb
-    rm virtualbox-5.0_5.0.8-103449~Ubuntu~trusty_amd64.deb
-  elif [[ $(cat /etc/lsb-release | grep ' 12') != '' ]]; then
-    wget http://download.virtualbox.org/virtualbox/5.0.8/virtualbox-5.0_5.0.8-103449~Ubuntu~precise_amd64.deb
-    dpkg -i virtualbox-5.0_5.0.8-103449~Ubuntu~precise_amd64.deb
-    rm virtualbox-5.0_5.0.8-103449~Ubuntu~precise_amd64.deb
+  if [[ $(dpkg -l | egrep '^i' | egrep 'virtualbox-5.0\s+5.0.8') == '' ]]; then
+    if [[ $(cat /etc/lsb-release | grep ' 15\.10') != '' ]]; then
+      wget http://download.virtualbox.org/virtualbox/5.0.8/virtualbox-5.0_5.0.8-103449~Ubuntu~wily_amd64.deb
+      dpkg -i virtualbox-5.0_5.0.8-103449~Ubuntu~wily_amd64.deb
+      rm virtualbox-5.0_5.0.8-103449~Ubuntu~wily_amd64.deb
+    elif [[ $(cat /etc/lsb-release | egrep ' (14|15\.04)') != '' ]]; then
+      wget http://download.virtualbox.org/virtualbox/5.0.8/virtualbox-5.0_5.0.8-103449~Ubuntu~trusty_amd64.deb
+      dpkg -i virtualbox-5.0_5.0.8-103449~Ubuntu~trusty_amd64.deb
+      rm virtualbox-5.0_5.0.8-103449~Ubuntu~trusty_amd64.deb
+    else
+      wget http://download.virtualbox.org/virtualbox/5.0.8/virtualbox-5.0_5.0.8-103449~Ubuntu~precise_amd64.deb
+      dpkg -i virtualbox-5.0_5.0.8-103449~Ubuntu~precise_amd64.deb
+      rm virtualbox-5.0_5.0.8-103449~Ubuntu~precise_amd64.deb
+    fi
   else
-    echo 'ERROR: No appropriate VirtualBox package found'
+    echo 'VirtualBox is already installed'
   fi
 
   # Install Ruby
-  # T.B.D.
+  apt-add-repository -y ppa:brightbox/ruby-ng
+  apt-get -y update
+  apt-get -y install ruby2.2
   gem install bundler -v 1.10.5
 
 elif [[ "$OSTYPE" == "darwin"* ]]; then
